EthProxy uses uint128 math now

+++ contracts/interfaces/IDealer.sol
@@ -5,13 +5,13 @@ import "./IYDai.sol";
 
 interface IDealer {
     function series(uint256) external returns (IYDai);
-    function systemPosted(bytes32) external returns (uint256);
-    function systemDebtYDai(bytes32, uint256) external returns (uint256);
-    function posted(bytes32, address) external view returns (uint256);
-    function totalDebtDai(bytes32, address) external returns (uint256);
+    function systemPosted(bytes32) external returns (uint128);
+    function systemDebtYDai(bytes32, uint256) external returns (uint128);
+    function posted(bytes32, address) external view returns (uint128);
+    function totalDebtDai(bytes32, address) external returns (uint128);
     function isCollateralized(bytes32, address) external returns (bool);
-    function grab(bytes32, address, uint256, uint256) external;
+    function grab(bytes32, address, uint128, uint128) external;
     function shutdown() external;
-    function post(bytes32, address, address, uint256) external;
-    function withdraw(bytes32, address, address, uint256) external;
+    function post(bytes32, address, address, uint128) external;
+    function withdraw(bytes32, address, address, uint128) external;
 }
